Performance: Optimize Regex compilation in markdown processing
