+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/CreateVoiceActivity.kt
@@ -508,14 +508,18 @@ class CreateVoiceActivity : BaseActivity() {
         var processed = markdown.replace("\n", "  \n")
         if (pendingImages.isNotEmpty()) {
             val pendingByFileName = pendingImages.values.associateBy { it.fileName }
-            processed = IMAGE_MARKDOWN_REGEX.replace(processed) { matchResult ->
-                val path = matchResult.groupValues.getOrNull(1).orEmpty()
-                val pending = pendingByFileName[path]
-                if (pending != null) {
-                    val prefix = matchResult.value.substringBeforeLast("(") + "("
-                    "$prefix${pending.file.toURI()})"
-                } else {
-                    matchResult.value
+            if (pendingByFileName.isNotEmpty()) {
+                val globalPattern = Regex("(!\\[[^\\]]*\\]\\()(.*?)(\\))")
+                processed = globalPattern.replace(processed) { matchResult ->
+                    val path = matchResult.groupValues.getOrNull(2).orEmpty()
+                    val pending = pendingByFileName[path]
+                    if (pending != null) {
+                        val prefix = matchResult.groupValues.getOrNull(1).orEmpty()
+                        val suffix = matchResult.groupValues.getOrNull(3).orEmpty()
+                        "$prefix${pending.file.toURI()}$suffix"
+                    } else {
+                        matchResult.value
+                    }
                 }
             }
         }
@@ -1500,7 +1504,6 @@ class CreateVoiceActivity : BaseActivity() {
     }
 
     companion object {
-        private val IMAGE_MARKDOWN_REGEX = Regex("!\\[[^\\]]*\\]\\(([^)]+)\\)")
         private const val PREVIEW_DEBOUNCE_MS = 150L
         private const val MAX_HEADING_LEVEL = 6
         private val NUMBERED_LIST_REGEX = Regex("^(\\d+)\\.\\s*(.*)$")
